+++ packages/snaps-controllers/package.json
@@ -23,7 +23,7 @@
     "test:prepare": "yarn mkdirp test/fixtures && ./scripts/generate-fixtures.sh",
     "test": "rimraf coverage && yarn test:prepare && jest && yarn test:browser && yarn posttest",
     "posttest": "ts-node scripts/coverage.ts && rimraf coverage/jest coverage/wdio",
-    "test:browser": "wdio run wdio.config.ts",
+    "test:browser": "wdio run wdio.config.js",
     "test:ci": "yarn test",
     "build": "yarn build:source && yarn build:types",
     "build:source": "yarn build:esm && yarn build:cjs",
