+++ packages/snaps-controllers/tsconfig.build.json
@@ -12,7 +12,6 @@
     "./src/**/test-utils",
     "./src/**/__mocks__",
     "./src/**/__snapshots__",
-    "wdio.config.ts",
     "scripts"
   ],
   "references": [
