+++ packages/snaps-controllers/tsconfig.json
@@ -3,7 +3,7 @@
   "compilerOptions": {
     "baseUrl": "./"
   },
-  "include": ["./src", "wdio.config.ts", "scripts"],
+  "include": ["./src", "scripts"],
   "references": [
     { "path": "../snaps-execution-environments" },
     { "path": "../rpc-methods" },
