+++ packages/snaps-execution-environments/tsconfig.build.json
@@ -13,7 +13,6 @@
     "./src/**/test-utils",
     "./src/**/__mocks__",
     "./src/**/__snapshots__",
-    "wdio.config.ts",
     "scripts"
   ],
   "references": [
