+++ packages/snaps-execution-environments/tsconfig.json
@@ -5,13 +5,7 @@
     "typeRoots": ["../../node_modules/@types", "./node_modules/@types"],
     "allowJs": true
   },
-  "include": [
-    "./src",
-    "./src/openrpc.json",
-    "webpack.config.js",
-    "wdio.config.ts",
-    "scripts"
-  ],
+  "include": ["./src", "./src/openrpc.json", "webpack.config.js", "scripts"],
   "references": [
     {
       "path": "../rpc-methods"
