+++ packages/snaps-execution-environments/wdio.config.js
@@ -1,13 +1,19 @@
-import { NodeGlobalsPolyfillPlugin } from '@esbuild-plugins/node-globals-polyfill';
-import { NodeModulesPolyfillPlugin } from '@esbuild-plugins/node-modules-polyfill';
-import type { Options } from '@wdio/types';
-import { resolve } from 'path';
-import tsconfigPaths from 'vite-tsconfig-paths';
+/* eslint-disable import/unambiguous, @typescript-eslint/no-require-imports, @typescript-eslint/no-var-requires */
 
+const {
+  NodeGlobalsPolyfillPlugin,
+} = require('@esbuild-plugins/node-globals-polyfill');
+const {
+  NodeModulesPolyfillPlugin,
+} = require('@esbuild-plugins/node-modules-polyfill');
+const { resolve } = require('path');
+const { default: tsconfigPaths } = require('vite-tsconfig-paths');
+
+// eslint-disable-next-line n/no-process-env
 const IS_CI = Boolean(process.env.CI);
 const MAX_WORKERS = IS_CI ? 1 : 5;
 
-export const config: Options.Testrunner = {
+const config = {
   runner: [
     'browser',
     {
@@ -20,7 +26,6 @@ export const config: Options.Testrunner = {
               NodeModulesPolyfillPlugin(),
               NodeGlobalsPolyfillPlugin({
                 buffer: true,
-                process: true,
               }),
             ],
           },
@@ -64,39 +69,38 @@ export const config: Options.Testrunner = {
     [
       'static-server',
       {
-        port: 4567,
+        port: 4568,
+        middleware: [
+          {
+            mount: '/worker/executor/',
+
+            // This is a workaround to add the `Access-Control-Allow-Origin`
+            // header to the worker bundle, which is required because it is
+            // loaded from a different origin than the test page.
+            middleware: (_, response) => {
+              response.setHeader('Access-Control-Allow-Origin', '*');
+
+              response.type('application/javascript');
+              response.sendFile(
+                resolve(
+                  __dirname,
+                  './dist/browserify/worker-executor/bundle.js',
+                ),
+              );
+            },
+          },
+        ],
         folders: [
           // The iframe execution service bundle.
           {
             mount: '/',
-            path: resolve(
-              __dirname,
-              '../snaps-execution-environments/dist/browserify/iframe',
-            ),
+            path: resolve(__dirname, './dist/browserify/iframe'),
           },
 
           // The web worker execution service bundle.
           {
             mount: '/worker/executor',
-            path: resolve(
-              __dirname,
-              '../snaps-execution-environments/dist/browserify/worker-executor',
-            ),
-          },
-
-          // The web worker pool.
-          {
-            mount: '/worker/pool',
-            path: resolve(
-              __dirname,
-              '../snaps-execution-environments/dist/browserify/worker-pool',
-            ),
-          },
-
-          // A test page used for testing the sandboxing.
-          {
-            mount: '/test/sandbox',
-            path: resolve(__dirname, './src/services/iframe/test'),
+            path: resolve(__dirname, './dist/browserify/worker-executor'),
           },
         ],
       },
@@ -119,3 +123,5 @@ export const config: Options.Testrunner = {
     timeout: 120000,
   },
 };
+
+module.exports.config = config;
