+++ packages/snaps-utils/tsconfig.build.json
@@ -12,7 +12,6 @@
     "./src/**/__mocks__",
     "./src/**/__snapshots__",
     "./src/eval-worker.js",
-    "wdio.config.ts",
     "scripts"
   ],
   "references": [
