+++ packages/snaps-utils/tsconfig.json
@@ -3,6 +3,6 @@
   "compilerOptions": {
     "baseUrl": "./"
   },
-  "include": ["./src", "wdio.config.ts", "scripts"],
+  "include": ["./src", "scripts"],
   "references": [{ "path": "../snaps-ui" }]
 }
